shift bite

+++ labs/memory_bound/data_packing/init.cpp
@@ -19,10 +19,10 @@
 S create_entry(int first_value, int second_value) {
   S entry;
 
-  entry.i = static_cast<short>(first_value);
+  entry.i = first_value;
   entry.s = static_cast<short>(second_value);
-  entry.l = static_cast<int>(first_value * second_value);
-  entry.d = static_cast<float>(first_value) / maxRandom;
+  entry.l = static_cast<long long>(first_value * second_value);
+  entry.d = static_cast<double>(first_value) / maxRandom;
   entry.b = first_value < second_value;
 
   return entry;
